evaluacion requerimientos pauta lll

+++ app/src/main/java/com/example/minefarms/data/entity/FavoriteEntity.kt
@@ -12,7 +12,7 @@ data class FavoriteEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
     val userId: Long,
-    val farmId: Int, // ID de la granja predefinida
+    val farmId: Long, // ID de la granja (ahora Long para consistencia con FarmEntity)
     val isLiked: Boolean = false,
     val isSaved: Boolean = false,
     val addedAt: Long = System.currentTimeMillis()
